prj/djn/requirements.txt
asgiref==3.3.1
boto3==1.16.56
botocore==1.19.56
cent==3.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coverage==5.3.1
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.1.5
django-elasticsearch-dsl==7.1.4
django-rest-elasticsearch==0.4.2
django-storages==1.11.1
djangorestframework==3.12.2
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
factory-boy==3.2.0
Faker==5.6.5
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
selenium==3.141.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
